Simplify Color type and implement ColorHistogram
Color type was unnecessarily complicated, duplicating too much of the
the standard library's image/color.RGBA type, so now it is an alias for
this type with an additional method to render it as a string. The Color
type is not even really needed, now that I think about it, but I'll
remove it later.

The real value of this commit is the ColorHistogram, which now can
generate a slice of top-N most used colors out of all the pixels that
have been added to it.